# help/dist/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "Minter Explorer api"
  },
  "paths": {
    "/api/v1/blocks": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "Get list of blocks",
        "operationId": "getBlocks",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of blocks",
            "content": {
              "application-json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BlockResource"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blocks/{height}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "Get block detail",
        "operationId": "getBlock",
        "parameters": [
          {
            "name": "height",
            "in": "path",
            "description": "Block height",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of blocks",
            "content": {
              "application-json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BlockResource"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Block not found",
            "content": {
              "application-json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer",
                          "example": 404
                        },
                        "message": {
                          "type": "string",
                          "example": "Block not found."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/coins": {
      "get": {
        "tags": [
          "Coins"
        ],
        "summary": "Get list of coins",
        "operationId": "getCoins",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "Coin symbol",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of coins",
            "content": {
              "application-json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CoinResource"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlockResource": {
        "properties": {
          "height": {
            "description": "Block height",
            "type": "integer",
            "example": 34863
          },
          "size": {
            "description": "Block size",
            "type": "integer",
            "example": 253
          },
          "txCount": {
            "description": "Transaction count",
            "type": "integer",
            "example": 25
          },
          "blockTime": {
            "description": "Block time",
            "type": "integer",
            "example": 34863
          },
          "timestamp": {
            "description": "Block height",
            "type": "integer",
            "example": 518
          },
          "reward": {
            "description": "Block reward",
            "type": "string",
            "example": "333"
          },
          "hash": {
            "description": "Block hash",
            "type": "string",
            "example": "ITHA0XQPS5XTOY6NTTOOJX5VAVZQHTRIZJDHQKKUPUQDNX7JYVDE4UO5LFFCWSVA"
          }
        },
        "type": "object"
      },
      "CoinResource": {
        "properties": {
          "crr": {
            "description": "Coin crr",
            "type": "integer",
            "example": 10
          },
          "volume": {
            "description": "Coin volume",
            "type": "string",
            "example": "45656534.3546"
          },
          "reserve_balance": {
            "description": "Coin reserve balance",
            "type": "string",
            "example": "10.356"
          },
          "name": {
            "description": "Coin name",
            "type": "string",
            "example": "My super coin"
          },
          "symbol": {
            "description": "Coin symbol",
            "type": "string",
            "example": "MNT"
          }
        },
        "type": "object"
      }
    }
  }
}